+++ ChangeLog.md
@@ -1,5 +1,6 @@
 - Adds a `Cut` effect which adds committed choice to nondeterminism.
 - Fixes the table of contents links in the README.
+- Adds an example of using `NonDet`, `Cut`, and a character parser effect to define parsers.
 
 # 0.1.1.0